qt, util: Split out qt/guifileutil module
