+++ src/Makefile.qt.include
@@ -119,6 +119,7 @@ BITCOIN_QT_H = \
   qt/csvmodelwriter.h \
   qt/editaddressdialog.h \
   qt/guiconstants.h \
+  qt/guifileutil.h \
   qt/guiutil.h \
   qt/intro.h \
   qt/macdockiconhandler.h \
@@ -219,6 +220,7 @@ BITCOIN_QT_BASE_CPP = \
   qt/bitcoinunits.cpp \
   qt/clientmodel.cpp \
   qt/csvmodelwriter.cpp \
+  qt/guifileutil.cpp \
   qt/guiutil.cpp \
   qt/intro.cpp \
   qt/modaloverlay.cpp \
