+++ src/qt/guiutil.h
@@ -6,7 +6,6 @@
 #define BITCOIN_QT_GUIUTIL_H
 
 #include <amount.h>
-#include <fs.h>
 
 #include <QEvent>
 #include <QHeaderView>
@@ -89,38 +88,6 @@ namespace GUIUtil
 
     void setClipboard(const QString& str);
 
-    /**
-     * Determine default data directory for operating system.
-     */
-    QString getDefaultDataDirectory();
-
-    /** Get save filename, mimics QFileDialog::getSaveFileName, except that it appends a default suffix
-        when no suffix is provided by the user.
-
-      @param[in] parent  Parent window (or 0)
-      @param[in] caption Window caption (or empty, for default)
-      @param[in] dir     Starting directory (or empty, to default to documents directory)
-      @param[in] filter  Filter specification such as "Comma Separated Files (*.csv)"
-      @param[out] selectedSuffixOut  Pointer to return the suffix (file type) that was selected (or 0).
-                  Can be useful when choosing the save file format based on suffix.
-     */
-    QString getSaveFileName(QWidget *parent, const QString &caption, const QString &dir,
-        const QString &filter,
-        QString *selectedSuffixOut);
-
-    /** Get open filename, convenience wrapper for QFileDialog::getOpenFileName.
-
-      @param[in] parent  Parent window (or 0)
-      @param[in] caption Window caption (or empty, for default)
-      @param[in] dir     Starting directory (or empty, to default to documents directory)
-      @param[in] filter  Filter specification such as "Comma Separated Files (*.csv)"
-      @param[out] selectedSuffixOut  Pointer to return the suffix (file type) that was selected (or 0).
-                  Can be useful when choosing the save file format based on suffix.
-     */
-    QString getOpenFileName(QWidget *parent, const QString &caption, const QString &dir,
-        const QString &filter,
-        QString *selectedSuffixOut);
-
     /** Get connection type to call object slot in GUI thread with invokeMethod. The call will be blocking.
 
        @returns If called from the GUI thread, return a Qt::DirectConnection.
@@ -136,12 +103,6 @@ namespace GUIUtil
 
     // Set shortcut to close window
     void handleCloseWindowShortcut(QWidget* w);
-
-    // Open debug.log
-    void openDebugLogfile();
-
-    // Open the config file
-    bool openBitcoinConf();
 
     /** Qt event filter that intercepts ToolTipChange events, and replaces the tooltip with a rich text
       representation if needed. This assures that Qt can word-wrap long tooltip messages.
@@ -214,15 +175,6 @@ namespace GUIUtil
             void on_sectionResized(int logicalIndex, int oldSize, int newSize);
             void on_geometriesChanged();
     };
-
-    bool GetStartOnSystemStartup();
-    bool SetStartOnSystemStartup(bool fAutoStart);
-
-    /* Convert QString to OS specific boost path through UTF-8 */
-    fs::path qstringToBoostPath(const QString &path);
-
-    /* Convert OS specific boost path to QString through UTF-8 */
-    QString boostPathToQString(const fs::path &path);
 
     /* Convert seconds into a QString with days, hours, mins, secs */
     QString formatDurationStr(int secs);
