+++ src/qt/paymentserver.cpp
@@ -9,6 +9,7 @@
 #include <qt/paymentserver.h>
 
 #include <qt/bitcoinunits.h>
+#include <qt/guifileutil.h>
 #include <qt/guiutil.h>
 #include <qt/optionsmodel.h>
